# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface GithubUser {
  id: number;
  login: string;
  avatarUrl: string;
}

export interface GithubAuthStatus {
  connected: boolean;
  user: GithubUser | null;
}

export interface GithubRepository {
  id: number;
  name: string;
  fullName: string;
  owner: string;
  /** @nullable */
  description: string | null;
  private: boolean;
  defaultBranch: string;
  updatedAt: string;
}

export interface GithubBuildInput {
  /** @minLength 1 */
  owner: string;
  /** @minLength 1 */
  repo: string;
  /** @minLength 1 */
  branch: string;
}

export type GithubBuildEventLevel = typeof GithubBuildEventLevel[keyof typeof GithubBuildEventLevel];


export const GithubBuildEventLevel = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  error: 'error',
} as const;

export interface GithubBuildEvent {
  id: number;
  message: string;
  level: GithubBuildEventLevel;
  createdAt: string;
}

export type GithubBuildStatus = typeof GithubBuildStatus[keyof typeof GithubBuildStatus];


export const GithubBuildStatus = {
  queued: 'queued',
  in_progress: 'in_progress',
  success: 'success',
  failure: 'failure',
  cancelled: 'cancelled',
} as const;

export interface GithubBuild {
  id: string;
  owner: string;
  repo: string;
  branch: string;
  status: GithubBuildStatus;
  attempt: number;
  maxAttempts: number;
  /** @nullable */
  runId: number | null;
  /** @nullable */
  artifactId: number | null;
  events: GithubBuildEvent[];
  createdAt: string;
  updatedAt: string;
}