# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /github/builds:
    post:
      operationId: startGithubBuild
      tags: [github]
      summary: Validate a repository and dispatch an Android build
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubBuildInput"
      responses:
        "201":
          description: Build dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubBuild"
        "400":
          description: Repository structure is incomplete
        "401":
          description: GitHub connection required
  /github/builds/{buildId}:
    get:
      operationId: getGithubBuild
      tags: [github]
      summary: Get build progress and event history
      parameters:
        - name: buildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Build status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubBuild"
        "404":
          description: Build not found
...
components:
  schemas:
...
    GithubBuildInput:
      type: object
      properties:
        owner:
          type: string
          minLength: 1
        repo:
          type: string
          minLength: 1
        branch:
          type: string
          minLength: 1
      required: [owner, repo, branch]
    GithubBuildEvent:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        level:
          type: string
          enum: [info, success, warning, error]
        createdAt:
          type: string
      required: [id, message, level, createdAt]
    GithubBuild:
      type: object
      properties:
        id:
          type: string
        owner:
          type: string
        repo:
          type: string
        branch:
          type: string
        status:
          type: string
          enum: [queued, in_progress, success, failure, cancelled]
        attempt:
          type: integer
        maxAttempts:
          type: integer
        runId:
          type: ["integer", "null"]
        artifactId:
          type: ["integer", "null"]
        events:
          type: array
          items:
            $ref: "#/components/schemas/GithubBuildEvent"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        [id, owner, repo, branch, status, attempt, maxAttempts, runId, artifactId, events, createdAt, updatedAt]