# lib/api-zod/src/generated/types/githubBuild.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubBuildEvent } from './githubBuildEvent';
import type { GithubBuildStatus } from './githubBuildStatus';

export interface GithubBuild {
  id: string;
  owner: string;
  repo: string;
  branch: string;
  status: GithubBuildStatus;
  attempt: number;
  maxAttempts: number;
  /** @nullable */
  runId: number | null;
  /** @nullable */
  artifactId: number | null;
  events: GithubBuildEvent[];
  createdAt: string;
  updatedAt: string;
}