# lib/api-zod/src/generated/types/githubBuildEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubBuildEventLevel } from './githubBuildEventLevel';

export interface GithubBuildEvent {
  id: number;
  message: string;
  level: GithubBuildEventLevel;
  createdAt: string;
}