# lib/api-zod/src/generated/types/githubBuildInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubBuildInput {
  /** @minLength 1 */
  owner: string;
  /** @minLength 1 */
  repo: string;
  /** @minLength 1 */
  branch: string;
}